+++ src/App.jsx
@@ -1,13 +1,13 @@
 import "./App.css";
 import Header from "./components/shared/Header/Header.component";
-import InventorySection from "./components/sections/InventorySection.component";
+import MainContainer from "./components/ui/MainContainer.component";
 import Footer from "./components/shared/Footer.component";
 
 function App() {
   return (
     <>
       <Header />
-      <InventorySection />
+      <MainContainer />
       <Footer />
     </>
   );